Support projection and ordering
The special field '_only' can be used to specify a list of field
names to fetch. Defaults to * (fetch everything).

The field '_order' can be used to add an ORDER BY clause. It can be a
single field name (for ascending) or a Field-Dir pair (where Dir is
either asc or desc).